Issue asd804 launch way

+++ model/src/main/java/com/emc/mongoose/model/impl/data/DataCorruptionException.java
@@ -8,7 +8,7 @@ extends DataVerificationException {
 	//
 	public final byte expected, actual;
 	//
-	public DataCorruptionException(final int relOffset, final byte expected, final byte actual) {
+	public DataCorruptionException(final long relOffset, final byte expected, final byte actual) {
 		super(relOffset);
 		this.expected = expected;
 		this.actual = actual;
